+++ web/src/modules/centre/pages/CentreDashboardWorksheetsMonthlyWorksheetPage.vue
@@ -14,66 +14,28 @@
       color="primary"
       size="small"
       class="mt-3"
+      :loading="isInitializing"
       @click="initializeWorksheetsForFiscalYear"
       >Add worksheets for {{ fiscalYear }}</v-btn
     >
   </div>
-  <div
+  <FundingSubmissionLineJsonEditSheet
     v-else
+    :funding-submission-line-json-id="fundingSubmissionLineJsonId"
     class="ma-4"
-  >
-    <v-btn
-      color="primary"
-      class="float-right"
-      @click="saveFundingSubmissionLineJson"
-      :loading="isSaving"
-      >Save</v-btn
-    >
-
-    <h2 class="mb-3">{{ dateName }} {{ calendarYear }}</h2>
-    <v-btn
-      v-if="dateName == FIRST_FISCAL_MONTH_NAME"
-      :loading="isReplicatingEstimates"
-      color="yg_sun"
-      class="float-right mb-3"
-      size="small"
-      @click="replicateEstimatesForward"
-    >
-      <v-icon>mdi-content-copy</v-icon> Replicate Estimates
-    </v-btn>
-
-    <div
-      v-for="({ sectionName, lines }, sectionIndex) in sections"
-      :key="`${sectionName}-${sectionIndex}`"
-      style="clear: both"
-    >
-      <h4>{{ sectionName }}</h4>
-
-      <SectionTable
-        ref="sectionTables"
-        :lines="lines"
-        @line-changed="propagateUpdatesAsNeeded(sectionIndex, $event)"
-      />
-    </div>
-  </div>
+  />
 </template>
 
 <script lang="ts" setup>
 import { computed, ref } from "vue"
-import { groupBy, isEmpty, upperFirst } from "lodash"
-import { DateTime } from "luxon"
+import { isEmpty, upperFirst } from "lodash"
 
 import centresApi from "@/api/centres-api"
 import { FiscalPeriodMonths } from "@/api/fiscal-periods-api"
-import fundingSubmissionLineJsonsApi from "@/api/funding-submission-line-jsons-api"
 import { useNotificationStore } from "@/store/NotificationStore"
-import useFundingSubmissionLineJsons, {
-  type FundingLineValue,
-} from "@/use/use-funding-submission-line-jsons"
+import useFundingSubmissionLineJsons from "@/use/use-funding-submission-line-jsons"
 
-import SectionTable from "@/modules/centre/components/centre-dashboard-worksheets-tab-monthly-worksheet-tab/SectionTable.vue"
-
-const FIRST_FISCAL_MONTH_NAME = "April"
+import FundingSubmissionLineJsonEditSheet from "@/components/funding-submission-line-jsons/FundingSubmissionLineJsonEditSheet.vue"
 
 const notificationStore = useNotificationStore()
 
@@ -101,28 +63,11 @@ const {
 
 const fundingSubmissionLineJson = computed(() => fundingSubmissionLineJsons.value[0])
 const fundingSubmissionLineJsonId = computed(() => fundingSubmissionLineJson.value?.id)
-const dateName = computed(() => fundingSubmissionLineJson.value?.dateName)
 
-const isSaving = ref(false)
-const isReplicatingEstimates = ref(false)
-const calendarYear = computed(() =>
-  DateTime.fromISO(fundingSubmissionLineJson.value?.dateStart).toFormat("yyyy")
-)
-const sections = computed<{ sectionName: string; lines: FundingLineValue[] }[]>(() => {
-  if (isEmpty(fundingSubmissionLineJson.value)) {
-    return []
-  }
-
-  const lines = fundingSubmissionLineJson.value.lines
-  const sectionGroups = groupBy(lines, "sectionName")
-  return Object.entries(sectionGroups).map(([sectionName, lines]) => {
-    return { sectionName, lines }
-  })
-})
-const sectionTables = ref<InstanceType<typeof SectionTable>[]>([])
+const isInitializing = ref(false)
 
 async function initializeWorksheetsForFiscalYear() {
-  isSaving.value = true
+  isInitializing.value = true
   try {
     await centresApi.fiscalYear.create(props.centreId, fiscalYear.value)
     notificationStore.notify({
@@ -131,58 +76,7 @@ async function initializeWorksheetsForFiscalYear() {
     })
     await refreshFundingSubmissionLineJsons()
   } finally {
-    isSaving.value = false
-  }
-}
-
-async function saveFundingSubmissionLineJson() {
-  isSaving.value = true
-  try {
-    const lines = sections.value.flatMap((section) => section.lines)
-    await fundingSubmissionLineJsonsApi.update(fundingSubmissionLineJsonId.value, { lines })
-  } catch (error) {
-    console.error(error)
-    notificationStore.notify({
-      text: `Failed to save worksheet: ${error}`,
-      variant: "error",
-    })
-  } finally {
-    isSaving.value = false
-  }
-}
-
-async function replicateEstimatesForward() {
-  isReplicatingEstimates.value = true
-  try {
-    await saveFundingSubmissionLineJson()
-    await fundingSubmissionLineJsonsApi.replicateEstimates(fundingSubmissionLineJsonId.value)
-  } catch (error) {
-    notificationStore.notify({
-      text: `Failed to replicate estimates: ${error}`,
-      variant: "error",
-    })
-  } finally {
-    isReplicatingEstimates.value = false
-  }
-}
-
-function propagateUpdatesAsNeeded(
-  sectionIndex: number,
-  { line, lineIndex }: { line: FundingLineValue; lineIndex: number }
-) {
-  // Bind section 1 to sections 2 and 3
-  // When you update the values in section 1, it will propagated the values to section 2 and 3
-  if (sectionIndex === 0) {
-    const section1Line = sections.value[1].lines[lineIndex]
-    section1Line.estimatedChildOccupancyRate = line.estimatedChildOccupancyRate
-    section1Line.actualChildOccupancyRate = line.actualChildOccupancyRate
-
-    sectionTables.value[1].refreshLineTotals(section1Line)
-
-    const section2Line = sections.value[2].lines[lineIndex]
-    section2Line.estimatedChildOccupancyRate = line.estimatedChildOccupancyRate
-    section2Line.actualChildOccupancyRate = line.actualChildOccupancyRate
-    sectionTables.value[2].refreshLineTotals(section2Line)
+    isInitializing.value = false
   }
 }
 </script>
